requirements.txt
openai
python-dotenv
langchain
langchain-openai
langchain-core
langchain-community
langchain-text-splitters
faiss-cpu
numpy
gdown
fastapi
uvicorn
websockets
python-multipart
httpx
